x_do: XDo#display is optional, fix focused_window
The `display` field in an `XDo` instance is now populated by
default from `ENV["DISPLAY"]`, which can be `nil`. This means that
`display` itself can potentially be `nil`.

This commit also fixes the `XDo#focused_window` method, which was
broken because `out` cannot be used on the same variable more than
once.